refactor: move cache utils out of hook
refs: FILES-185 FILES-396
